requirements.txt
feedparser
requests
python-dateutil
python-dotenv
anthropic
